+++ tests/services/test_savedsearch.py
@@ -1,6 +1,7 @@
 from tests.conftest import create_mock_response
 
 from py42 import settings
+from py42.sdk.queries.fileevents.file_event_query import FileEventQuery
 from py42.services.fileevent import FileEventService
 from py42.services.savedsearch import SavedSearchService
 
@@ -114,4 +115,64 @@ class TestSavedSearchService:
         assert (
             mock_connection.get.call_args[0][0]
             == "/forensic-search/queryservice/api/v1/saved/test-id"
+        )
+
+    # V2 TESTS
+    def test_get_calls_with_v2_uri_if_v2_settings_enabled(
+        self, mock_connection, mocker
+    ):
+        import py42.settings
+
+        py42.settings.use_v2_file_event_data = True
+        file_event_service = FileEventService(mock_connection)
+        saved_search_service = SavedSearchService(mock_connection, file_event_service)
+        saved_search_service.get()
+        py42.settings.use_v2_file_event_data = False
+        mock_connection.get.assert_called_once_with(
+            "/forensic-search/queryservice/api/v2/saved"
+        )
+
+    def test_get_by_id_calls_v2_uri_if_v2_settings_enabled(self, mock_connection):
+        import py42.settings
+
+        py42.settings.use_v2_file_event_data = True
+        file_event_service = FileEventService(mock_connection)
+        saved_search_service = SavedSearchService(mock_connection, file_event_service)
+        saved_search_service.get_by_id("test-id")
+        py42.settings.use_v2_file_event_data = False
+        mock_connection.get.assert_called_once_with(
+            "/forensic-search/queryservice/api/v2/saved/test-id"
+        )
+
+    def test_get_query_builds_v2_query_if_v2_settings_enabled(
+        self, mock_connection, mocker
+    ):
+        import py42.settings
+
+        py42.settings.use_v2_file_event_data = True
+        response = create_mock_response(mocker, SAVED_SEARCH_GET_RESPONSE)
+        mock_connection.post.return_value = response
+        file_event_service = FileEventService(mock_connection)
+        saved_search_service = SavedSearchService(mock_connection, file_event_service)
+        query = saved_search_service.get_query("test-id")
+        py42.settings.use_v2_file_event_data = False
+        assert query.sort_key == "event.id"
+
+    def test_execute_calls_search_with_v2_uri_if_v2_settings_enabled(
+        self, mock_connection, mocker
+    ):
+        import py42.settings
+
+        py42.settings.use_v2_file_event_data = True
+
+        response = create_mock_response(mocker, SAVED_SEARCH_GET_RESPONSE)
+        mock_connection.post.return_value = response
+
+        file_event_service = FileEventService(mock_connection)
+        saved_search_service = SavedSearchService(mock_connection, file_event_service)
+        saved_search_service.execute("test-id")
+        expected_query = FileEventQuery.from_dict({"groups": []})
+        py42.settings.use_v2_file_event_data = False
+        mock_connection.post.assert_called_once_with(
+            "/forensic-search/queryservice/api/v2/fileevent", json=dict(expected_query)
         )